PS1 fix and stopped .emacs whitespace removal
Changed PS1 for non-dumb prompts to appropriately work with emacs
shortcuts. E.g. C-a should go to the beginning of a line, but the way
bash was interpreting PS1 was wrong so C-a would go three characters
after the point where it should be. The input command to the shell was
then effectively shifted (basically a whole mess). By moving around
the color settings, this has been remedied. Also, I stopped doing
whitesapce removal whenever emacs saves. This created some annoying
behavior when you would save a file and expect to stay on the same
location in a blank line.